+++ instrumentation/opentelemetry-instrumentation-sqlalchemy/tests/test_sqlcommenter.py
@@ -46,10 +46,11 @@ class TestSqlalchemyInstrumentationWithSQLCommenter(TestBase):
             engine=engine,
             tracer_provider=self.tracer_provider,
             enable_commenter=True,
+            commenter_options={"db_framework": False},
         )
         cnx = engine.connect()
         cnx.execute("SELECT  1;").fetchall()
         self.assertRegex(
             self.caplog.records[-2].getMessage(),
-            r"SELECT  1 /\*traceparent='\d{1,2}-[a-zA-Z0-9_]{32}-[a-zA-Z0-9_]{16}-\d{1,2}'\*/;",
+            r"SELECT  1 /\*db_driver='(.*)',traceparent='\d{1,2}-[a-zA-Z0-9_]{32}-[a-zA-Z0-9_]{16}-\d{1,2}'\*/;",
         )